pp/cache: add scheduling mechanisms to GC
This commit adds the GC mutex to `remove_client_if()` which is always
called during the GC process. Using a mutex protects against the
situation where a client is stopped when there are still inflight
requests.
